only questions of valid question pools are returned to the client
